cron: add task to run automatic followup and erase of neglected drafts
When a news draft was not updated since 5 months and 2 weeks,
a job sends an automatic followup email to redactors and a message
on the news board.

When a news draft was not updated since 6 months,
a job erases it automatically and sends the news content
to the author of the draft and to the moderation team.