display login name with username
To avoid issue with visual spoofing, we can add login aside the user
display name. This ensure we can always distinguish users, because login
names have to be unique.

To avoid to display too many information in the same line below the
title, the editors and moderator can be hidden with the `<details>` HTML
tag.

Details are hidden by default on content lists and they are shown on
content page.