modernize RUN commands in Containerfiles
As RUN commands on multiline is like a shell script, enable the bash
strict mode. So we can remove all the `&&` at the begin of lines.

When buildah (podman builder) recent release will be published in common
distributions (at least Debian stable), we will be able to use the
"heredoc" syntax to avoid all the `; \` at end of lines.